+++ axelor-helpdesk/src/main/java/com/axelor/apps/helpdesk/web/TicketController.java
@@ -18,7 +18,7 @@
 package com.axelor.apps.helpdesk.web;
 
 import com.axelor.apps.base.db.Timer;
-import com.axelor.apps.base.db.TimerState;
+import com.axelor.apps.base.db.repo.TimerRepository;
 import com.axelor.apps.base.service.app.AppBaseService;
 import com.axelor.apps.helpdesk.db.Ticket;
 import com.axelor.apps.helpdesk.service.TicketService;
@@ -147,7 +147,7 @@ public class TicketController {
       boolean hideStart = false;
       boolean hideCancel = true;
       if (timer != null) {
-        hideStart = timer.getState() == TimerState.STARTED;
+        hideStart = timer.getStatusSelect() == TimerRepository.TIMER_STARTED;
         hideCancel = timer.getTimerHistoryList().isEmpty();
       }