+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Remove unneeded required attribute for decimal in accounting situation
 - Rework TeamTask UI
 - Production processes: display status in grid view.
+- Rework Ticket UI.
 
 ## Bug Fixes
 - Check duplicate : Apply manual filter before duplicate check and allow to open form view.